List 17 Task #55 Added Disable/Enable key to SessionLog,DataLog and ErrorLog (Master,Biweekly)
