+++ web.config
@@ -160,7 +160,18 @@
     <add key="BehaviorSharePointRpt" value="http://sqltest/sites/DatasheetReports/_layouts/15/ReportServer/AdHocReportDesigner.aspx?list=%7B6E3DFEE8%2D00B2%2D48EF%2DAB36%2D0C18B279B701%7D&amp;ID=2"/>
     <add key="DataSheetSharePointRpt" value="http://sqltest/sites/DatasheetReports/_layouts/15/ReportServer/AdHocReportDesigner.aspx?list=%7B6E3DFEE8%2D00B2%2D48EF%2DAB36%2D0C18B279B701%7D&amp;ID=1"/>
   
-  </appSettings>
+			<!-- Session Log Enable/disable-->
+    		<add key="EnableBiSessionLog" value="true" />
+
+    		<!-- Error Log Enable/disable-->
+    		<add key="EnableBiErrorLog" value="true" />
+
+    		<!-- Score_Data Enable/disable-->
+    		<add key="EnableBiScore_Data" value="true" />
+    
+
+
+	</appSettings>
   <connectionStrings>
     <!--<add name="dbConnectionString" connectionString="Data Source=M2CLP003-PC\DHILOGICS;Initial Catalog=AppDataProdP1;User ID=sa;Password=sqladmin" providerName="System.Data.SqlClient"/>-->
     <!--<add name="dbConnectionString" connectionString="Data Source=M2CLP007;Initial Catalog=MelmarkDB;User ID=sa;Password=sqladmin" providerName="System.Data.SqlClient"/>-->
